Removed redundant s3Helper.go from logging, and consolidated with existing file in util/s3.
