+++ scripts/upload_version.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 # upload_version.sh - Uploads available update packages
+#           NOTE: Will only work if you have the required AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY vars set
 #
 # Syntax:   upload_version.sh <channel> <directory-with-packages>
 #
